requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
gunicorn==21.2.0

# 機械学習・データ処理
scipy==1.11.4
matplotlib==3.7.2
seaborn==0.12.2

# ファイル処理
openpyxl==3.1.2

# ログ・デバッグ
python-dateutil==2.8.2

# CORS・PWA対応
flask-cors==4.0.0

# 環境変数管理（必要に応じて）
python-dotenv==1.0.0

# 🔥 非同期処理追加
celery==5.3.4
redis==5.0.1
flower==2.0.1
psutil==5.9.6